requirements.txt
absl-py==0.12.0
albumentations==0.5.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
black==22.3.0
bleach==3.3.0
cachetools==4.2.1
certifi==2020.6.20
cffi==1.14.3
charset-normalizer==3.2.0
clang==5.0
click==8.1.3
cmake==3.27.4.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
faiss-gpu==1.7.0
filelock==3.3.0
fire==0.5.0
flake8==6.1.0
flatbuffers==1.12
fsspec==0.9.0
future==0.18.2
gast==0.3.3
google-auth==1.28.1
google-auth-oauthlib==0.4.4
grpcio==1.37.0
h5py==2.10.0
huggingface-hub==0.0.19
idna==3.4
imageio==2.9.0
imgaug==0.4.0
install==1.3.4
ipykernel==5.5.3
ipympl==0.7.0
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
keras-efficientnet-v2==1.1.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lightgbm==3.2.1
lit==16.0.6
llvmlite==0.37.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.7.0
mistune==0.8.4
mpmath==1.3.0
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
notebook==6.3.0
numba==0.54.0
numpy==1.18.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.3.0.36
opencv-python-headless==4.5.1.48
opt-einsum==3.3.0
packaging==23.0
pandas==1.1.1
pandocfilters==1.4.3
parso==0.7.1
pathlib==1.0.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
platformdirs==2.5.2
plotly==5.3.1
prometheus-client==0.10.1
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.11.0
pycparser==2.20
pydegensac==0.1.2
pydicom==2.0.0
pyflakes==3.1.0
Pygments==2.7.1
pynndescent==0.5.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==1.2.7
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.10.8
requests==2.31.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
safetensors==0.3.3
scikit-image==0.17.2
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
Shapely==1.7.0
sympy==1.12
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==2.6.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
thop==0.0.31.post2005241907
tifffile==2020.8.13
timm==0.9.7
tokenizers==0.10.3
tomli==2.0.1
torch==2.0.1
torchmetrics==0.2.0
torchvision==0.7.0
tornado==6.1
tqdm==4.48.2
traitlets==5.0.5
transformers==4.11.3
triton==2.0.0
typing-extensions==4.2.0
umap-learn==0.5.1
urllib3==2.0.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
